% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_bagpipes.R
\name{check_bagpipes}
\alias{check_bagpipes}
\title{Check whether the Bagpipes Python backend can be imported}
\usage{
check_bagpipes(python = Sys.which("python3"))
}
\arguments{
\item{python}{Python executable.}
}
\value{
Invisibly returns \code{TRUE} if Bagpipes can be imported.
}
\description{
Check whether the Bagpipes Python backend can be imported
}
