% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_set.R
\name{jwst_nircam_filter_set}
\alias{jwst_nircam_filter_set}
\title{JWST/NIRCam filter set bundled with sagui}
\usage{
jwst_nircam_filter_set(filters = NULL)
}
\arguments{
\item{filters}{Optional NIRCam filters/order.}
}
\value{
A \code{sagui_filter_set} object.
}
\description{
JWST/NIRCam filter set bundled with sagui
}
